Refactor mobx stores object into ./DataStores/index.js
